requirements.txt
fastapi
passlib[bcrypt]
sqlalchemy
python-dotenv
pydantic
uvicorn
python-jose
python-multipart